满立减列表

+++ app/src/main/java/cn/ifhu/mershop/bean/DiscountBean.java
@@ -13,8 +13,8 @@ public class DiscountBean {
      */
     private int xianshi_id;
     private String xianshi_name;
-    private int start_time;
-    private int end_time;
+    private long start_time;
+    private long end_time;
     private int state;
     private int lower_limit;
 
@@ -42,19 +42,19 @@ public class DiscountBean {
         this.xianshi_name = xianshi_name;
     }
 
-    public int getStart_time() {
+    public long getStart_time() {
         return start_time;
     }
 
-    public void setStart_time(int start_time) {
+    public void setStart_time(long start_time) {
         this.start_time = start_time;
     }
 
-    public int getEnd_time() {
+    public long getEnd_time() {
         return end_time;
     }
 
-    public void setEnd_time(int end_time) {
+    public void setEnd_time(long end_time) {
         this.end_time = end_time;
     }
